REST update

+++ src/main/resources/db/migration/V1_0__tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS Company (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     logo_url VARCHAR(255),
     address VARCHAR(255),
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS Company (
 );
 
 CREATE TABLE IF NOT EXISTS Extras (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     remote BOOLEAN DEFAULT false,
     flexibel BOOLEAN DEFAULT false,
     sign_up BOOLEAN DEFAULT false,
@@ -19,10 +19,10 @@ CREATE TABLE IF NOT EXISTS Extras (
 );
 
 CREATE TABLE IF NOT EXISTS Jobs (
-    id INT PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT PRIMARY KEY AUTO_INCREMENT,
     title VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
-    company_id INT NOT NULL,
+    company_id BIGINT NOT NULL,
     anzahl INT DEFAULT 1,
     status ENUM('Praktikum', 'Lehrstelle', 'Job'),
     pensum ENUM('20', '40', '60', '80', '100'),
